WIP Serverless Function and Layer to AWS SAM Template
This commit introduces a new Makefile for building an AWS Lambda layer and updates the AWS SAM template to define an AWS Serverless Function and Layer. The build process includes creating necessary directories and installing dependencies specified in `requirements.txt`. This setup ensures the function and layer are properly configured for deployment on AWS with Python 3.10 runtime.

This is to adapt that shared lambda layer so that it can be built in a way that is more compatible with AWS SAM and locally invoking with a debugger for  better productivity